requirements.txt
numpy
faiss-cpu
sentence-transformers
datasets
scikit-learn
pandas
dotenv